+++ src/Enums/Certificate/Action.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace SebastianSulinski\LaravelForgeSdk\Enums\Certificate;
+
+enum Action: string
+{
+    case Enable = 'enable';
+    case Disable = 'disable';
+}